SimpliPlay 4.6 (further updated) source
This update fixes a bug in the original (no longer available for download) 4.6 that mapped gesture detection to a non-Android view-type component (resulting in a runtime error of mapping to an element of incorrect type). The mapping has been changed to a view type element, which will fix the error and allow long tapping to work properly.